refactor: set feature through window function
