+++ packages/shared/src/lib/cookie.ts
@@ -1,0 +1,47 @@
+import { FeatureValue } from './featureManagement';
+import { checkIsPreviewDeployment } from './links';
+
+export const COOKIE_FEATURES_KEY = 'preview:featuresFlags';
+
+export const getCookieObject = (): Record<string, unknown> => {
+  if (typeof document === 'undefined') {
+    return {};
+  }
+
+  const cookie = {};
+  document.cookie.split(';').forEach((el) => {
+    const [key, value] = el.split('=');
+    cookie[key.trim()] = decodeURIComponent(value);
+  });
+
+  return cookie;
+};
+
+export const getCookieFeatureFlags = (): Record<string, unknown> => {
+  const cookie = getCookieObject();
+  const value = cookie[COOKIE_FEATURES_KEY] as string;
+
+  if (!value) {
+    return {};
+  }
+
+  try {
+    const features = JSON.parse(value);
+    delete features.flags;
+    return features;
+  } catch (err) {
+    return {};
+  }
+};
+
+export const isPreviewDeployment = checkIsPreviewDeployment();
+
+if (isPreviewDeployment) {
+  const setFeature = (key: string, value: FeatureValue) => {
+    const features = getCookieFeatureFlags();
+    features[key] = value;
+    document.cookie = `${COOKIE_FEATURES_KEY}=${JSON.stringify(features)}`;
+  };
+
+  globalThis.setFeature = setFeature;
+}
